Refactor to prevent collision of OSGi init with maven init
