+++ tycho-bundles/org.eclipse.tycho.embedder.shared/src/main/java/org/eclipse/tycho/Requirements.java
@@ -1,10 +1,8 @@
-package org.eclipse.tycho.p2.resolver.facade;
+package org.eclipse.tycho;
 
 import java.util.List;
 
-import org.eclipse.tycho.IDependencyMetadata;
 import org.eclipse.tycho.IDependencyMetadata.DependencyMetadataType;
-import org.eclipse.tycho.RequiredCapability;
 
 public interface Requirements {
