--- tycho-bundles/org.eclipse.tycho.p2.resolver.impl/OSGI-INF/requirements.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.tycho.p2.resolver.P2Requirements">
-   <implementation class="org.eclipse.tycho.p2.resolver.P2Requirements"/>
-   <service>
-      <provide interface="org.eclipse.tycho.p2.resolver.facade.Requirements"/>
-   </service>
-</scr:component>
